# openapi_v3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "angie",
    "version": "1.0.0",
    "license": {
      "name": "SEE LICENSE IN LICENSE"
    },
    "description": "Specification JSONs: [v2](/api-spec/v2), [v3](/api-spec/v3).\n\nAngie's "
  },
  "paths": {
...
    "/configuration/model/{code}": {
      "get": {
        "summary": "/configuration/model/{code}",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "configuration"
        ]
      }
    },
    "/configuration/model/{code}/data": {
      "get": {
        "summary": "/configuration/model/{code}/data",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "configuration"
        ]
      },
      "post": {
        "summary": "/configuration/model/{code}/data",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "configuration"
        ]
      }
    },
    "/configuration/model/{code}/data/{id}": {
      "get": {
        "summary": "/configuration/model/{code}/data/{id}",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "configuration"
        ]
      },
      "put": {
        "summary": "/configuration/model/{code}/data/{id}",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "configuration"
        ]
      },
      "delete": {
        "summary": "/configuration/model/{code}/data/{id}",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {},
        "tags": [
          "configuration"
        ]
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://localhost:3105"
    }
  ],
  "components": {
    "securitySchemes": {
      "authorization": {
        "name": "authorization",
        "in": "header",
        "type": "apiKey"
      }
    }
  }
}